fix: refactor helper class for web api tests
