+++ tests/CNAB.TestHelpers/CNAB.TestHelpers.csproj
@@ -7,6 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Moq" Version="4.20.72" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\..\src\CNAB.Application\CNAB.Application.csproj" />
     <ProjectReference Include="..\..\src\CNAB.Domain\CNAB.Domain.csproj" />
     <ProjectReference Include="..\..\src\CNAB.Infra.Data\CNAB.Infra.Data.csproj" />
